Refactor - replacing JSON with YAML + Wrapping Sacred
