--- cfg/full_cfg_deepisp.json
@@ -1,53 +1,0 @@
-{
-  "train_cfg":  {
-                "model": "cfg/model_cfg/deepisp_reg_conv_ll15_hl3.cfg",
-                // why no comments?
-                "max_epochs": 5000,
-                // optimizer parms
-                "optimizer_type": "Adam",
-                "initial_lr": 5e-5,
-                "momentum": 0.9,
-                "weight_decay": 0, //edit this
-                // scheduler params
-                "scheduler_type": "steps",
-                "step_size": 0,
-                "gamma": 0,
-                // train with half precision
-                "fp16_train": true,
-                //criterion
-                "loss_fn": "DeepISPLoss"
-                },
-
-
-  "handlers_cfg": {
-                  "dirname": "exp_deepisp_deep_16",
-                  "filename_prefix": "deep_isp",
-                  "save_interval": 1,
-                  "n_saved": 10,
-                  "multi_gpu": true,
-                  "patience": 1000,
-                  "score_function_name": "L1",
-                  "score_is_loss": true
-                  },
-
-  "dataloader_cfg": {
-                    "train_csv_file": "dataset/Sony_train_list.txt",
-                    "eval_csv_file": "dataset/Sony_val_list.txt",
-                    "root_dir": "/home/dsteam/Desktop",
-                    "black_level": 512,
-                    "stack_bayer": true,
-                    "subset": 1,
-                    "crop_size": 512,
-                    "batch_size": 2,
-                    "shuffle": true,
-                    "num_workers": 2,
-                    "pin_memory": true
-                    },
-
-  "prune_cfg" : {
-                "model": "cfg/model_cfg/orig_u_net.cfg",
-                "pruning_iterations": 9,
-                "filters_to_remove": 10, // in percents %
-                "recovery_epochs": 5
-                }
-}
