--- cfg/full_cfg_deepisp.yml
@@ -1,43 +1,0 @@
----
-TRAIN:
-    Model: cfg/model_cfg/deepisp_reg_conv_ll15_hl3.cfg
-    MaxEpoch: 5000
-    Optimizer:
-        OptimizerType: Adam
-        LearningRate: 5e-5
-        Momentum: 0.9
-        WeightDecay: 0
-    Scheduler:
-        SchedulerType: steps
-        StepSize: 0
-        Gamma: 0
-        FP16Train: True # train with half precision
-        LossFunction: DeepISPLoss # criterion
-
-HANDLERS:
-    Dirname: exp_deepisp_deep_16
-    FilenamePrefix: exp_deepisp_deep_16
-    SaveInterval: 1
-    NumSaved: 10
-    MultiGPU: True
-    Patience: 1000
-    ScoreFunctionName: L1
-    ScoreIsLoss: True
-    Resume: True
-
-
-DATALOADER:
-    TrainPath: dataset/Sony_train_list.txt
-    EvalPath: dataset/Sony_val_list.txt
-    RootDir: /home/dsteam/Desktop
-    SID:
-      BlackLevel: 512
-      StackBayer: True
-    Subset: 1
-    CropSize: 512
-    BatchSize: 2 # see https://github.com/cchen156/Learning-to-See-in-the-Dark/issues/16
-    Shuffle: True
-    NumWorkers: 0
-    PinMemory: True
-
-
