--- cfg/full_cfg_sid.json
@@ -1,47 +1,0 @@
-{
-  "train_cfg":  {
-                "model": "cfg/model_cfg/orig_u_net.cfg",
-                // why no comments?
-                "max_epochs": 4000,
-                // optimizer parms
-                "optimizer_type": "Adam",
-                "initial_lr": 1e-4,
-                "momentum": 0.9,
-                "weight_decay": 0, //edit this
-                // scheduler params
-                "scheduler_type": "steps",
-                "step_size": 2000,
-                "gamma": 0.1,
-                // train with half precision
-                "fp16_train": true,
-                //criterion
-                "loss_fn": "L1"
-                },
-
-
-  "handlers_cfg": {
-                  "dirname": "sid_vanilla",
-                  "filename_prefix": "sid_vanilla",
-                  "save_interval": 1,
-                  "n_saved": 10,
-                  "multi_gpu": true,
-                  "patience": 1000,
-                  "score_function_name": "L1",
-                  "score_is_loss": true,
-                  "resume": true
-                  },
-
-  "dataloader_cfg": {
-                    "train_csv_file": "dataset/Sony_train_list.txt",
-                    "eval_csv_file": "dataset/Sony_val_list.txt",
-                    "root_dir": "/home/dsteam/Desktop",
-                    "black_level": 512,
-                    "stack_bayer": true,
-                    "subset": 1,
-                    "crop_size": 512,
-                    "batch_size": 2, // see https://github.com/cchen156/Learning-to-See-in-the-Dark/issues/16
-                    "shuffle": true,
-                    "num_workers": 0,
-                    "pin_memory": true
-                    }
-}
